agent-app/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
pymysql
langchain
langchain-openai
langchain-core
